rename delete modal package for clarity
