+++ packages/composer-playground/src/app/delete-confirm/delete-confirm.component.scss
