+++ packages/composer-playground/src/app/delete-confirm/delete-confirm.component.spec.ts
@@ -5,7 +5,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
-import { DeleteComponent } from './delete.component';
+import { DeleteComponent } from './delete-confirm.component';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 describe('DeleteComponent', () => {
