+++ packages/composer-playground/src/app/delete-confirm/delete-confirm.component.ts
@@ -4,8 +4,8 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
     selector: 'confirm',
-    templateUrl: './delete.component.html',
-    styleUrls: ['./delete.component.scss'.toString()]
+    templateUrl: './delete-confirm.component.html',
+    styleUrls: ['./delete-confirm.component.scss'.toString()]
 })
 export class DeleteComponent {
     @Input() delete;
