+++ packages/composer-playground/src/app/delete-confirm/index.ts
@@ -1,0 +1,1 @@
+export * from './delete-confirm.component';
