--- packages/composer-playground/src/app/delete/index.ts
@@ -1,1 +1,0 @@
-export * from './delete.component';
